+++ butane/base/util/file.go
@@ -47,7 +47,15 @@ func ReadLocalFile(configPath, filesDir string) ([]byte, error) {
 	if err := EnsurePathWithinFilesDir(filePath, filesDir); err != nil {
 		return nil, err
 	}
-	return os.ReadFile(filePath)
+
+	file, err := os.Open(filePath)
+	if err != nil {
+		return nil, err
+	}
+	defer file.Close()
+
+	// TODO: keep old branch, if based on version?
+	return GomplateReadLocalFile(file)
 }
 
 // CheckForDecimalMode fails if the specified mode appears to have been